Move APU to an Audio Worklet.
I'm not sure if this is an improvement at all or not. Maybe latency is
better? Needs more testing.

Also bundled HTML loaded from file:// works in all browsers except
Chrome, so we'll need to polyfill there using ye olde
ScriptProcessorNode.
