+++ runtimes/web/src/apu.js
@@ -1,218 +1,37 @@
-const SAMPLE_RATE = 44100;
-const MAX_VOLUME = 0.25;
-const BUFFER_SIZE = 512; // Might need to bump this to 1024
-
-class Channel {
-    /** Starting frequency. */
-    freq1 = 0;
-
-    /** Ending frequency, or zero for no frequency transition. */
-    freq2 = 0;
-
-    /** Time the tone was started. */
-    startTime = 0;
-
-    /** Time at the end of the attack period. */
-    attackTime = 0;
-
-    /** Time at the end of the decay period. */
-    decayTime = 0;
-
-    /** Time at the end of the sustain period. */
-    sustainTime = 0;
-
-    /** Time the tone should end. */
-    releaseTime = 0;
-
-    /** Sustain volume level. */
-    volume = 0;
-
-    /** Used for time tracking. */
-    phase = 0;
-
-    /** Duty cycle for pulse channels. */
-    pulseDutyCycle = 0;
-
-    /** Noise generation state. */
-    noiseSeed = 0x0001;
-
-    /** The last generated random number, either -1 or 1. */
-    noiseLastRandom = 0;
-}
-
-function lerp (value1, value2, t) {
-    return value1 + t * (value2 - value1);
-}
-
-function polyblep (phase, phaseInc) {
-    if (phase < phaseInc) {
-        const t = phase / phaseInc;
-        return t+t - t*t;
-    } else if (phase > 1 - phaseInc) {
-        const t = (phase - (1 - phaseInc)) / phaseInc;
-        return 1 - (t+t - t*t);
-    } else {
-        return 1;
-    }
-}
+import workletRawSource from "../dist/apu-worklet.js?raw"; // RELEASE
+// import workletRawSource from "./apu-worklet.js?raw"; // DEBUG
 
 export class APU {
-    constructor () {
-        this.time = 0;
-        this.channels = new Array(4);
-        for (let ii = 0; ii < 4; ++ii) {
-            this.channels[ii] = new Channel();
-        }
-
-        const ctx = new (window.AudioContext || window.webkitAudioContext)({
-            sampleRate: SAMPLE_RATE,
+    async init () {
+        const audioCtx = new (window.AudioContext || window.webkitAudioContext)({
+            sampleRate: 44100, // must match SAMPLE_RATE in worklet
         });
-        this.ctx = ctx;
+        this.audioCtx = audioCtx;
 
-        const scriptNode = ctx.createScriptProcessor(BUFFER_SIZE, 0, 1);
-        scriptNode.addEventListener("audioprocess", event => {
-            const output = event.outputBuffer.getChannelData(0);
+        const blob = new Blob([workletRawSource], {type: "application/javascript"});
+        const url = URL.createObjectURL(blob);
+        await audioCtx.audioWorklet.addModule(url);
 
-            for (let ii = 0; ii < BUFFER_SIZE; ++ii, ++this.time) {
-                let sum = 0;
-
-                for (let channelIdx = 0; channelIdx < 4; ++channelIdx) {
-                    const channel = this.channels[channelIdx];
-                    if (this.time < channel.releaseTime) {
-                        const freq = this.getCurrentFrequency(channel);
-                        const volume = this.getCurrentVolume(channel);
-                        let sample;
-
-                        if (channelIdx == 3) {
-                            // Noise channel
-                            channel.phase += freq * freq / (1000000/44100 * SAMPLE_RATE);
-                            while (channel.phase > 0) {
-                                channel.phase--;
-                                let noiseSeed = channel.noiseSeed;
-                                noiseSeed ^= noiseSeed >> 7;
-                                noiseSeed ^= noiseSeed << 9;
-                                noiseSeed ^= noiseSeed >> 13;
-                                channel.noiseSeed = noiseSeed;
-                                channel.noiseLastRandom = ((noiseSeed & 0x1) << 1) - 1;
-                            }
-                            sample = volume * channel.noiseLastRandom;
-
-                        } else {
-                            const phaseInc = freq / SAMPLE_RATE;
-                            let phase = channel.phase + phaseInc;
-
-                            if (phase >= 1) {
-                                phase--;
-                            }
-                            channel.phase = phase;
-
-                            if (channelIdx == 2) {
-                                // Triangle channel
-                                sample = volume * Math.abs(2*phase - 1);
-
-                            } else {
-                                // Pulse channel
-                                let dutyPhase, dutyPhaseInc, multiplier;
-
-                                // Map duty to 0->1
-                                const pulseDutyCycle = channel.pulseDutyCycle;
-                                if (phase < pulseDutyCycle) {
-                                    dutyPhase = phase / pulseDutyCycle;
-                                    dutyPhaseInc = phaseInc / pulseDutyCycle;
-                                    multiplier = volume;
-                                } else {
-                                    dutyPhase = (phase - pulseDutyCycle) / (1 - pulseDutyCycle);
-                                    dutyPhaseInc = phaseInc / (1 - pulseDutyCycle);
-                                    multiplier = -volume;
-                                }
-                                sample = multiplier * polyblep(dutyPhase, dutyPhaseInc);
-                            }
-                        }
-
-                        sum += sample;
-                    }
-                }
-
-                output[ii] = sum;
-            }
-        });
-        scriptNode.connect(ctx.destination);
-    }
-
-    ramp (value1, value2, time1, time2) {
-        const t = (this.time - time1) / (time2 - time1);
-        return lerp(value1, value2, t);
-    }
-
-    getCurrentFrequency (channel) {
-        if (channel.freq2 > 0) {
-            return this.ramp(channel.freq1, channel.freq2, channel.startTime, channel.releaseTime);
-        } else {
-            return channel.freq1;
-        }
-    }
-
-    getCurrentVolume (channel) {
-        const time = this.time;
-        if (time > channel.sustainTime) {
-            return this.ramp(channel.volume, 0, channel.sustainTime, channel.releaseTime);
-        } else if (time > channel.decayTime) {
-            return channel.volume;
-        } else if (time > channel.attackTime) {
-            return this.ramp(MAX_VOLUME, channel.volume, channel.attackTime, channel.decayTime);
-        } else if (channel.startTime != channel.attackTime) {
-            return this.ramp(0, MAX_VOLUME, channel.startTime, channel.attackTime);
-        } else {
-            return channel.volume;
-        }
+        const workletNode = new AudioWorkletNode(audioCtx, "wasm4-apu");
+        workletNode.connect(audioCtx.destination);
+        this.workletNode = workletNode;
     }
 
     tone (frequency, duration, volume, flags) {
-        const freq1 = frequency & 0xffff;
-        const freq2 = (frequency >> 16) & 0xffff;
+        this.workletNode.port.postMessage([frequency, duration, volume, flags]);
+    }
 
-        const sustain = (duration & 0xff);
-        const release = ((duration >> 8) & 0xff);
-        const decay = ((duration >> 16) & 0xff);
-        const attack = ((duration >> 24) & 0xff);
-
-        const channelIdx = flags & 0x3;
-        const mode = (flags >> 2) & 0x3;
-
-        const channel = this.channels[channelIdx];
-
-        // Restart the phase if this channel wasn't already playing
-        if (this.time > channel.releaseTime) {
-            channel.phase = (channelIdx == 2) ? 0.5 : 0;
+    unlockAudio () {
+        const audioCtx = this.audioCtx;
+        if (audioCtx.state == "suspended") {
+            audioCtx.resume();
         }
+    }
 
-        channel.freq1 = freq1;
-        channel.freq2 = freq2;
-        channel.startTime = this.time;
-        channel.attackTime = channel.startTime + ((SAMPLE_RATE*attack/60) >>> 0);
-        channel.decayTime = channel.attackTime + ((SAMPLE_RATE*decay/60) >>> 0);
-        channel.sustainTime = channel.decayTime + ((SAMPLE_RATE*sustain/60) >>> 0);
-        channel.releaseTime = channel.sustainTime + ((SAMPLE_RATE*release/60) >>> 0);
-        channel.volume = MAX_VOLUME * volume/100;
-
-        if (channelIdx == 0 || channelIdx == 1) {
-            switch (mode) {
-            case 0:
-                channel.pulseDutyCycle = 0.125;
-                break;
-            case 1: case 3: default:
-                channel.pulseDutyCycle = 0.25;
-                break;
-            case 2:
-                channel.pulseDutyCycle = 0.5;
-                break;
-            }
-
-        } else if (channelIdx == 2) {
-            // For the triangle channel, prevent popping on hard stops by adding a 1 ms release
-            if (release == 0) {
-                channel.releaseTime += (SAMPLE_RATE/1000) >>> 0;
-            }
+    pauseAudio () {
+        const audioCtx = this.audioCtx;
+        if (audioCtx.state == "running") {
+            audioCtx.suspend();
         }
     }
 }
